pyproject.toml
[build-system]
requires = [
  'setuptools>=61.0',
  'recommonmark',
]
build-backend = 'setuptools.build_meta'

[project]
name = "Orange-Spectroscopy"
description = "Extends Orange to handle spectral and hyperspectral analysis."
authors = [
    { name="Canadian Light Source, Biolab UL, Soleil, Elettra", email="marko.toplak@gmail.com"}
]
version = "0.9.1"
readme = {file = "README.pypi", content-type = "text/markdown"}
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
]
keywords = ["orange3 add-on", "spectroscopy", "infrared"]
dependencies = [
    'numpy>=1.24.0',
    'Orange3>=3.38.0',
    'orange-canvas-core>=0.2.4',
    'orange-widget-base>=4.25.0',
    'scipy>=1.10.0,<1.17.0',
    'scikit-learn>=1.5.1',
    'spectral>=0.22.3,!=0.23',
    'serverfiles>=0.2',
    'AnyQt>=0.2.0',
    'pyqtgraph>=0.13.1',
    'colorcet',
    'h5py',
    'extranormal3 >=0.0.3',
    'renishawWiRE>=0.1.8',
    'pillow>=9.2.0',
    'lmfit>=1.3.3',
    'bottleneck',
    'pebble',
    'agilent-format>=0.4.5',
    'pySNOM>=0.2.0',
]

[project.optional-dependencies]
doc = [
    "sphinx",
    "recommonmark",
    "sphinx_rtd_theme",
]
test = [
    "coverage",
]
dev = [
    "pre-commit",
    "ruff",
]

[project.urls]
Homepage = "https://github.com/Quasars/orange-spectroscopy"
Issues = "https://github.com/Quasars/orange-spectroscopy/issues"

[project.entry-points."orange3.addon"]
spectroscopy = "orangecontrib.spectroscopy"

[project.entry-points."orange.widgets.tutorials"]
"Spectroscopy Examples" = "orangecontrib.spectroscopy.tutorials"

[project.entry-points."orange.widgets"]
"Spectroscopy" = "orangecontrib.spectroscopy.widgets"

[project.entry-points."orange.canvas.help"]
html-index = "orangecontrib.spectroscopy.widgets:WIDGET_HELP_PATH"

[tool.setuptools]
namespace-packages = ["orangecontrib"]

[tool.setuptools.packages.find]
include = ["orangecontrib*"]

[tool.ruff]
force-exclude = true

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B"]
ignore = ["E741", "B904"]

[tool.ruff.format]
quote-style = "preserve"